Improiving version provider and bug fixing
