+++ cobigen-cli/pom.xml
@@ -4,7 +4,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.devonfw.cobigen</groupId>
   <artifactId>cli</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
+  <version>1.0.0-SNAPSHOT</version>
   <name>cobigen-cli</name>
 
   <properties>
@@ -19,12 +19,12 @@
     <dependency>
       <groupId>com.devonfw.cobigen</groupId>
       <artifactId>core</artifactId>
-      <version>5.3.4</version>
+      <version>5.3.5</version>
     </dependency>
     <dependency>
       <groupId>com.devonfw.cobigen</groupId>
       <artifactId>core-api</artifactId>
-      <version>5.3.4</version>
+      <version>5.3.5</version>
     </dependency>
     <dependency>
       <groupId>info.picocli</groupId>
