+++ cobigen-cli/src/main/java/com/devonfw/cobigen/cli/commands/CobiGenCommand.java
@@ -1,88 +1,23 @@
 package com.devonfw.cobigen.cli.commands;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Enumeration;
-import java.util.List;
-import java.util.Properties;
-
-import org.apache.maven.model.Dependency;
-import org.apache.maven.model.Model;
-import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
-import org.apache.maven.project.MavenProject;
-import org.apache.maven.shared.utils.xml.pull.XmlPullParserException;
-
-import com.devonfw.cobigen.cli.constants.MavenConstants;
 import com.devonfw.cobigen.cli.constants.MessagesConstants;
-import com.devonfw.cobigen.cli.utils.CobiGenUtils;
+import com.devonfw.cobigen.cli.utils.CobiGenVersionProvider;
 
-import net.sf.mmm.util.io.api.IoMode;
-import net.sf.mmm.util.io.api.RuntimeIoException;
-import net.sf.mmm.util.xml.base.XmlInvalidException;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.IVersionProvider;
-
 
 /**
  * This class defines the main CobiGen command
  */
 
-
 @Command(description = MessagesConstants.WELCOME_MESSAGE, name = "cobigen", aliases = { "cg" },
-    mixinStandardHelpOptions = true,  subcommands = { GenerateCommand.class },versionProvider = CobiGenCommand.PropertiesVersionProvider.class)
+    mixinStandardHelpOptions = true, subcommands = { GenerateCommand.class },
+    versionProvider = CobiGenVersionProvider.class)
 public class CobiGenCommand implements Runnable {
 
-	@Override
-	public void run() {
+    @Override
+    public void run() {
 
-		// Nothing to do here, this is the master command
-	}
-
-	/**
-	 * This class implement getVersion() and this method return the version of
-	 * plug-in
-	 */
-	static class PropertiesVersionProvider implements IVersionProvider {
-		@Override
-		public String[] getVersion() throws Exception {
-			List<String> versionProvider = new ArrayList<String>();
-			MavenXpp3Reader reader = new MavenXpp3Reader();
-			Model model = null;
-			File locationCLI = new File(
-					GenerateCommand.class.getProtectionDomain().getCodeSource().getLocation().toURI());
-			Path rootCLIPath = locationCLI.getParentFile().toPath();
-
-			File pomFile = new CobiGenUtils().extractArtificialPom(rootCLIPath);
-
-			if (pomFile.exists()) {
-				model = reader.read(new FileReader(pomFile));
-			}
-
-			List<Model> versionList = new ArrayList<Model>();
-			versionList.add(model);
-
-			List<Dependency> modelDependencies = model.getDependencies();
-
-			for (int i = 0; i < modelDependencies.size(); i++) {
-
-				versionProvider.add(" name:= " + modelDependencies.get(i).getArtifactId() + " version=  "
-						+ modelDependencies.get(i).getVersion());
-
-			}
-			return versionProvider.toArray(new String[versionProvider.size()]);
-		}
-
-	}
+        // Nothing to do here, this is the master command
+    }
 
 }
